[REFACTOR][RUNTIME] Phase out profiling.h heavy types, rename to timer.h
This PR removes the unused heavy profiling types (Profiler, Report,
MetricCollector, Duration/Percent/Count/Ratio, ShapeString,
DeviceWrapper, ProfileFunction) from runtime/profiling.h and renames
the header to runtime/timer.h. Only Timer, TimerNode, and
WrapTimeEvaluator are retained.

Main changes:

- Rename include/tvm/runtime/profiling.h → include/tvm/runtime/timer.h
- Rename src/runtime/profiling.cc → src/runtime/timer.cc
- Move WrapTimeEvaluator from tvm::runtime::profiling to tvm::runtime
- Rename FFI keys profiling.timer.<device> → runtime.timer.<device>
- Delete vm.profile() and VirtualMachineProfiler (TVM_VM_ENABLE_PROFILER)
- Delete json_runtime RunProfile/CanDebug/_debug dispatch path
- Delete CLML RunProfile override and dead MetricCollector declaration
- Delete python/tvm/runtime/profiling/ package entirely
